"local-fs.target" shouldn't depend on "zfs-mount.service"
The "zfs-mount.service" is responsible for mounting all ZFS datasets on
a Delphix appliance, including all domain0 mounts. On an appliance with
thousands of domain0 datasets (e.g. thousands of VDBs), it's expected
for the "zfs-mount.service" to take a relatively long time to start
(compared to most other bootup services). Thus, we don't want to cause
that to add latency to the entire bootup process, since lots (nearly
all) services have a dependency on "local-fs.target".

Thus, this change removes the "zfs-mount.service" dependency from the
"local-fs.target". This way, "local-fs.target" will start up, while the
domain0 mounts can proceed without blocking other bootup critical
services. This can safely be done because we rely on "/etc/fstab" for
mounting all of the root filesystem datasets, which do not use
"zfs-mount.service" and are already dependencies of "local-fs.target".

Thus, after this change, any service that depends on the root filesystem
datasets being mounted can depend on the "local-fs.target" (which is in
the default set of service dependencies). Any services that depends on
all domain0 datasets being mounted (e.g. NFS, iSCSI, etc.) can depend
explicitly on the "zfs-mount.service".

In addition to breaking the "local-fs" target's dependency on the
"zfs-mount" service, we also want to ensure the two services that import
zpools run after the "local-fs" target.

This is because those two services will grab the "spa_namespace_lock" in
the kernel, and may conflict with mounting the root datasets, causing
the root dataset mounts to be delayed (or even timeout).

Thus, in this change, we optimize for mounting the root datasets first,
and then all other datasets after (the "zfs-mount" service depends on
these two import services).

The assumption being made here is that all mounts required by the
"local-fs" target (e.g. all mounts done via "/etc/fstab") rely soley on
the root pool being imported; since only the root pool will be imported
when "local-fs.target" runs (it's imported by the initramfs/boot
process), and all other pools won't be imported until after the
"local-fs" target is finished.