Move tracing functions from "zfs" to "spl" module
This commit moves the tracing functions, that're generated by the
DTRACE_PROBE* macros, from the "zfs" module and into the "spl" module.
This is needed so that we can leverage all this infrastructure from the
"spl" module itself; i.e. to add tracing to the "spl" module.

When the DTRACE_PROBE* macros are used, they get expanded to create new
functions, and without this change these dynamically generated functions
are part of the "zfs" module.

Since the "spl" module does not depend on the "zfs" module, the use of
DTRACE_PROBE* in the "spl" module would result in undefined symbols
being used in the "spl" module. Specifically, DTRACE_PROBE* would turn
into a function call, and the function being called would be a symbol
only contained in the "zfs" module; which results in a linker and/or
runtime error.

In contrast, the "zfs" module does depend on the "spl" module. Thus,
these dynamically generated tracing functions can be contained in the
"spl" module, which then allows both the "zfs" and "spl" modules to take
advantage of the DTRACE_PROBE* infrastructure for tracing.

The changes being made in this commit do this restructuring, to move
these dynamically generated functions into the "spl" module.